+++ frontend/packages/data-portal/app/apiNormalization/browseAllDepositionsV2.ts
@@ -22,7 +22,7 @@ const remapV2Deposition = remapAPI<
   authors: (deposition) =>
     deposition.authors?.edges?.map((edge) => edge.node) ?? [],
   annotationCount: (deposition) =>
-    deposition.annotationsCount?.aggregate?.at(0)?.count ?? 0,
+    deposition.annotationsCount?.aggregate?.at(0)?.count ?? undefined,
   annotatedObjects: (deposition) =>
     deposition.objectNames?.aggregate?.reduce((acc, aggregate) => {
       const objectName = aggregate.groupBy?.objectName
@@ -41,6 +41,26 @@ const remapV2Deposition = remapAPI<
     [],
   acrossDatasets: (deposition) =>
     deposition.annotationDatasetCount?.aggregate?.length ?? 0,
+  tomogramsCount: (deposition) =>
+    deposition.tomogramsCount?.aggregate?.at(0)?.count ?? undefined,
+  totalImagingData: (deposition) =>
+    [
+      ...(deposition.framesFileSizes?.aggregate ?? []).map(
+        (aggregate) => aggregate.sum?.fileSize ?? 0,
+      ),
+
+      ...(deposition.tiltSeriesFileSizes?.aggregate ?? []).map(
+        (aggregate) =>
+          (aggregate.sum?.fileSizeMrc ?? 0) +
+          (aggregate.sum?.fileSizeOmezarr ?? 0),
+      ),
+
+      ...(deposition.tomogramFileSizes?.aggregate ?? []).map(
+        (aggregate) =>
+          (aggregate.sum?.fileSizeMrc ?? 0) +
+          (aggregate.sum?.fileSizeOmezarr ?? 0),
+      ),
+    ].reduce((total, fileSize) => total + fileSize, 0),
 } as const)
 
 export const remapV2BrowseAllDepositions = (