+++ frontend/packages/data-portal/app/utils/annotation.ts
@@ -1,0 +1,3 @@
+export function getAnnotationName(id: number, objectName: string | undefined) {
+  return `${id} ${objectName ?? ''}`
+}